Refactored DefaultTree tests. Fixes #846
